* commit 'adad54e2401b71e5d5193fcb038e035d3013d0f9':
  Added cannot in one word
  Modify Statement
  resources.py edited online with Bitbucket
  adding strip method to remove space
  resources.py edited online with Bitbucket
  bugFixed

+++ UserManagement/UserMgmt/src/db/db.py
@@ -163,6 +163,8 @@ def deleteSvcs(svc):
     elif check:
         return False
     try:
+        if not Services.objects(name=svc).first():
+            return False
         remSvcFrmRole('admin', read=[svc], write=[svc])
         Services.objects(name=svc).delete()
         return True
